[feat]: seongwon/ 파일분리, 폰트 설정, 색 통일, 3.5페이지 수정
