--- AppleSnack/MainCollectionViewCell.swift
@@ -1,12 +1,0 @@
-//
-//  MainCollectionViewCell.swift
-//  AppleSnack
-//
-//  Created by  on 2023/08/17.
-//
-
-import UIKit
-
-class MainCollectionViewCell: UICollectionViewCell {
-    @IBOutlet var mainCollectionViewCell: UILabel!
-}
