--- AppleSnack/Model(categorie)/Categorie+CoreDataClass.swift
@@ -1,15 +1,0 @@
-//
-//  Categorie+CoreDataClass.swift
-//  AppleSnack
-//
-//  Created by Macbook on 2023/08/19.
-//
-//
-
-import Foundation
-import CoreData
-
-@objc(Categorie)
-public class Categorie: NSManagedObject {
-
-}
