--- AppleSnack/Model(profile)/MyProfile+CoreDataClass.swift
@@ -1,15 +1,0 @@
-//
-//  MyProfile+CoreDataClass.swift
-//  AppleSnack
-//
-//  Created by Macbook on 2023/08/19.
-//
-//
-
-import Foundation
-import CoreData
-
-@objc(MyProfile)
-public class MyProfile: NSManagedObject {
-
-}
