--- AppleSnack/Model(snack)/MySnack+CoreDataClass.swift
@@ -1,15 +1,0 @@
-//
-//  MySnack+CoreDataClass.swift
-//  AppleSnack
-//
-//  Created by Macbook on 2023/08/19.
-//
-//
-
-import Foundation
-import CoreData
-
-@objc(MySnack)
-public class MySnack: NSManagedObject {
-
-}
