+++ AppleSnack/Views/MainCollectionViewCell.swift
@@ -1,0 +1,12 @@
+//
+//  MainCollectionViewCell.swift
+//  AppleSnack
+//
+//  Created by  on 2023/08/17.
+//
+
+import UIKit
+
+class MainCollectionViewCell: UICollectionViewCell {
+    @IBOutlet var mainCollectionViewCell: UILabel!
+}
